requirements.txt
# Usage: pip install -r requirements.txt

matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0,!=1.12.0
torchvision>=0.8.1,!=0.13.0
tqdm>=4.41.0
protobuf<4.21.3
tensorboard>=2.4.1
pandas>=1.1.4
seaborn>=0.11.0
Flask==2.1.2
gunicorn==20.1.0
pytesseract==0.3.9



